+++ src/renderer/PVESettings.tsx
@@ -40,17 +40,36 @@ const PVESettings = (props: IProps) => {
       return;
     }
 
-    setConfigValues({
+    const toSet: Record<string, boolean | string | number> = {
       recordRaids: config.recordRaids,
       minEncounterDuration: config.minEncounterDuration,
       minRaidDifficulty: config.minRaidDifficulty,
       recordDungeons: config.recordDungeons,
-      minKeystoneLevel: config.minKeystoneLevel,
       recordChallengeModes: config.recordChallengeModes,
       raidOverrun: config.raidOverrun,
       dungeonOverrun: config.dungeonOverrun,
       recordCurrentRaidEncountersOnly: config.recordCurrentRaidEncountersOnly,
-    });
+    };
+
+    // Only set these if they are valid values. We allow -1 set in the
+    // frontend to represent temporarily unset values in the Input fields.
+    if (config.minEncounterDuration >= 0) {
+      toSet.minEncounterDuration = config.minEncounterDuration;
+    }
+
+    if (config.raidOverrun >= 0) {
+      toSet.raidOverrun = config.raidOverrun;
+    }
+
+    if (config.minKeystoneLevel >= 0) {
+      toSet.minKeystoneLevel = config.minKeystoneLevel;
+    }
+
+    if (config.dungeonOverrun >= 0) {
+      toSet.dungeonOverrun = config.dungeonOverrun;
+    }
+
+    setConfigValues(toSet);
   }, [
     config.dungeonOverrun,
     config.minEncounterDuration,
@@ -152,6 +171,12 @@ const PVESettings = (props: IProps) => {
   const setMinEncounterDuration = (
     event: React.ChangeEvent<HTMLInputElement>,
   ) => {
+    if (!event.target.value) {
+      // Allow setting empty as midpoint.
+      setConfig((prev) => ({ ...prev, minEncounterDuration: -1 }));
+      return;
+    }
+
     const minEncounterDuration = parseInt(event.target.value, 10);
 
     if (Number.isNaN(minEncounterDuration)) {
@@ -189,7 +214,9 @@ const PVESettings = (props: IProps) => {
           </Tooltip>
         </Label>
         <Input
-          value={config.minEncounterDuration}
+          value={
+            config.minEncounterDuration >= 0 ? config.minEncounterDuration : ''
+          }
           name="minEncounterDuration"
           disabled={!config.recordRaids}
           onChange={setMinEncounterDuration}
@@ -251,6 +278,12 @@ const PVESettings = (props: IProps) => {
   };
 
   const setRaidOverrun = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (!event.target.value) {
+      // Allow setting empty as midpoint.
+      setConfig((prev) => ({ ...prev, raidOverrun: -1 }));
+      return;
+    }
+
     const raidOverrun = parseInt(event.target.value, 10);
 
     if (Number.isNaN(raidOverrun) || raidOverrun < 0 || raidOverrun > 60) {
@@ -286,7 +319,7 @@ const PVESettings = (props: IProps) => {
           </Tooltip>
         </Label>
         <Input
-          value={config.raidOverrun}
+          value={config.raidOverrun >= 0 ? config.raidOverrun : ''}
           name="raidOverrun"
           disabled={!config.recordRaids}
           onChange={setRaidOverrun}
@@ -297,6 +330,12 @@ const PVESettings = (props: IProps) => {
   };
 
   const setDungeonOverrun = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (!event.target.value) {
+      // Allow setting empty as midpoint.
+      setConfig((prev) => ({ ...prev, dungeonOverrun: -1 }));
+      return;
+    }
+
     const dungeonOverrun = parseInt(event.target.value, 10);
 
     if (
@@ -336,7 +375,7 @@ const PVESettings = (props: IProps) => {
           </Tooltip>
         </Label>
         <Input
-          value={config.dungeonOverrun}
+          value={config.dungeonOverrun >= 0 ? config.dungeonOverrun : ''}
           name="dungeonOverrun"
           disabled={!config.recordDungeons}
           onChange={setDungeonOverrun}
@@ -387,6 +426,12 @@ const PVESettings = (props: IProps) => {
   };
 
   const setMinKeystoneLevel = (event: React.ChangeEvent<HTMLInputElement>) => {
+    if (!event.target.value) {
+      // Allow setting empty as midpoint.
+      setConfig((prev) => ({ ...prev, minKeystoneLevel: -1 }));
+      return;
+    }
+
     const minKeystoneLevel = parseInt(event.target.value, 10);
 
     if (Number.isNaN(minKeystoneLevel)) {
@@ -421,7 +466,7 @@ const PVESettings = (props: IProps) => {
           </Tooltip>
         </Label>
         <Input
-          value={config.minKeystoneLevel}
+          value={config.minKeystoneLevel >= 0 ? config.minKeystoneLevel : ''}
           name="minKeystoneLevel"
           disabled={!config.recordDungeons}
           onChange={setMinKeystoneLevel}